pyproject.toml
[project]
name = "postgres-multi-mcp"
version = "0.4.0"
description = "PostgreSQL Multi-Database MCP Server - Manage multiple databases from a single MCP instance"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.25.0",
    "psycopg[binary]>=3.3.2",
    "humanize>=4.15.0",
    "pglast==7.11",
    "attrs>=25.4.0",
    "psycopg-pool>=3.3.0",
    "instructor>=1.14.4",
    "pyyaml>=6.0",
]
license = "mit"
license-files = ["LICENSE"]

[[project.authors]]
name = "psh4607"
email = "psh4607@example.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/postgres_mcp"]

[project.scripts]
postgres-mcp = "postgres_mcp:main"

# [tool.pyright]
# venvPath = "."
# venv = ".venv"

[tool.pytest.ini_options]
pythonpath = ["./src"]
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "docker>=7.1.0",
    "pyright==1.1.408",
    "pytest-asyncio>=1.3.0",
    "pytest>=9.0.2",
    "ruff==0.14.13",
]

[tool.black]
line-length = 150

[tool.ruff]
line-length = 150
target-version = "py39"
exclude = [".venv*"]

lint.select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "W",   # pycodestyle warnings
    # "C90", # mccabe complexity
    "N",   # pep8-naming
    # "D",   # pydocstyle
    "UP",  # pyupgrade
    "RUF"  # ruff-specific rules
]

# TODO: Remove these ignores when fixing #129 (code modernization)
lint.ignore = [
    "UP006",  # Use `list` instead of `List` for type annotations
    "UP035",  # Import from `collections.abc` instead of `typing`
    "UP045",  # Use `X | None` instead of `Optional[X]`
    "RUF059", # Unused unpacked variable
    "RUF100", # Unused noqa directive
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
force-single-line = true
known-first-party = ["postgres-mcp"]

[tool.pyright]
typeCheckingMode = "standard"
pythonVersion = "3.12"
reportMissingTypeStubs = false
# reportUnknownMemberType = true
# reportUnknownParameterType = true
# reportUnknownVariableType = true
# reportUnknownArgumentType = true
reportPrivateUsage = true
reportUntypedFunctionDecorator = true
reportMissingTypeArgument = true
reportInvalidTypeVarUse = true
reportUnnecessaryCast = true
# reportUnnecessaryComparison = true
